fix overflow

+++ test/yosupo/many_aplusb.test.cpp
@@ -4,7 +4,7 @@ int main() {
     int T;
     std::cin >> T;
     while (T--) {
-        int a, b;
+        ll a, b;
         std::cin >> a >> b;
         std::cout << a + b << std::endl;
     }
